requirements.txt
torch
pygithub
transformers
tiktoken
docker
pyautogen
matplotlib
#rubberband
yapf
icecream
lux
numpy
pandas
scikit-learn
scipy
python-dateutil
pip
datetime
jupyter
notebook==6.4.12
traitlets==5.9.0
jupyterthemes
ipykernel
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator
utils
seaborn
plotly
statsmodels
patsy
